member-phoneNum 추가 store-loc_detail 추가

+++ charlesLee_project/src/main/java/popup/dto/Member.java
@@ -22,7 +22,7 @@ public class Member {
 	private int memberType;
 	private String memberCompanyName;
 	private int memberCompanyNum;
-	private int memberPhoneNum;
+	private long memberPhoneNum;
 	private Date memberCreate;
 	private Date memberUpdate;
 }
